+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds an field to embeds for the suggested voice channel.
+
 ## [v14.3.0](https://github.com/lexicalunit/spellbot/releases/tag/v14.3.0) - 2025-01-01
 
 ### Changed